thermocouple: max31685: attempt to recover from faults
When `rtd_consecutive_fault_limit` is set to a value greater than 0,
all faults raised by the max31685 will be ignored, logged, and the
max31865 re-initialized using the original configuration up to the
fault limit. Once the max consecutive failures are reached, both the
firmware and host shut down (as it is today). Failures return the
last known good value.

The failure counter is maintained on the firmware side and passed along
with each thermocouple response. Upon a failure, the original
config_cmd is sent to re-initialize the max31865.

This is currently only implemented for the max31865.

Signed-off-by: Michael Rose <elementation@gmail.com>